Create test.ts
This is an example of using the wrapper. The selectedDate property is what will be updated by the Date Control and the binding engine. So, for example, if you have a contact entity with a birthdate, you would not have selectedDate here, but your entity.  Then you bind in the html to the entity.